// src/integrations/cms/generated/cms-schema.d.ts
export interface components {
    schemas: {
// ...
        ApiPromotionPromotionDocument: {
// ...
            content: unknown[];
// ...
        };
// ...
        ApiProductProductDocument: {
            /**
             * Format: uuid
             * @description The document ID, represented by a UUID
             */
            documentId: string;
            id: string | number;
            /** @description A string field */
            name: string;
            /** @description A UID field */
            slug: string;
            /** @description A text field */
            short_description?: string;
            /** @description A blocks field */
            description: unknown[];
            /**
             * @description An enum field
             * @enum {string}
             */
            product_type: "account" | "credit_card" | "loan" | "saving" | "insurance";
            /**
             * @description An enum field
             * @default active
             * @enum {string}
             */
            statusProduct: "active" | "inactive" | "coming_soon";
            /**
             * @description An enum field
             * @enum {string}
             */
            customerSegment?: "individual" | "corporate" | "priority";
            /**
             * @description An enum field
             * @enum {string}
             */
            targetAudience?: "Cá nhân" | "Doanh nghiệp" | "SME";
            /** @description An integer field */
            sortOrder?: number;
            /** @description A decimal field */
            interestRateMin?: number;
            /** @description A decimal field */
            interestRateMax?: number;
            /** @description A biginteger field */
            loanAmountMin?: string;
            /** @description A biginteger field */
            loanAmountMax?: string;
            /** @description A string field */
            tenor?: string;
            /** @description A string field */
            applyUrl?: string;
            /** @description A datetime field */
            createdAt?: string;
            /** @description A datetime field */
            updatedAt?: string;
            /**
             * @description A datetime field
             * @default 2026-06-04T06:49:03.468Z
             */
            publishedAt: string;
            /** @description A media field */
            thumbnail?: components["schemas"]["PluginUploadFileDocument"];
            /** @description A media field */
            main_banner?: components["schemas"]["PluginUploadFileDocument"];
            /** @description A component field */
            seo?: components["schemas"]["SharedSeoEntry"];
            /** @description A relational field */
            tags?: components["schemas"]["ApiTagTagDocument"][];
            /** @description A dynamic zone field */
            features?: unknown[];
            /** @description A dynamic zone field */
            benefits?: unknown[];
            /** @description A media field */
            images?: components["schemas"]["PluginUploadFileDocument"][];
            /** @description A media field */
            documents?: components["schemas"]["PluginUploadFileDocument"][];
            /** @description A media field */
            coverImage?: components["schemas"]["PluginUploadFileDocument"];
            /** @description A relational field */
            category?: components["schemas"]["ApiProductCategoryProductCategoryDocument"];
            /** @description A relational field */
            promotions?: components["schemas"]["ApiPromotionPromotionDocument"][];
            /** @description A relational field */
            faqs?: components["schemas"]["ApiFaqFaqDocument"][];
            /** @description A relational field */
            relatedProducts?: components["schemas"]["ApiProductProductDocument"][];
            /** @description A relational field */
            relatedArticles?: components["schemas"]["ApiArticleArticleDocument"][];
            /** @description A relational field */
            applicablePromotions?: components["schemas"]["ApiPromotionPromotionDocument"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    "faq/get/faqs_by_id": {
        parameters: {
            query?: {
                fields?: ("question" | "answer" | "slug" | "active" | "sortOrder" | "createdAt" | "updatedAt" | "publishedAt")[];
                populate?: "*" | ("category" | "product") | ("category" | "product")[];
                filters?: {
                    [key: string]: unknown;
                };
                sort?: ("question" | "answer" | "slug" | "active" | "sortOrder" | "createdAt" | "updatedAt" | "publishedAt") | ("question" | "answer" | "slug" | "active" | "sortOrder" | "createdAt" | "updatedAt" | "publishedAt")[] | {
                    [key: string]: "asc" | "desc";
                } | {
                    [key: string]: "asc" | "desc";
                }[];
                status?: "draft" | "published";
                hasPublishedVersion?: boolean | ("true" | "false");
            };
            header?: never;
            path: {
                id: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        data: {
                            /**
                             * Format: uuid
                             * @description The document ID, represented by a UUID
                             */
                            documentId: string;
                            id: string | number;
                            /** @description A string field */
                            question: string;
                            /** @description A blocks field */
                            answer: unknown[];
                            /** @description A UID field */
                            slug: string;
                            /**
                             * @description A boolean field
                             * @default true
                             */
                            active: boolean | null;
                            /** @description An integer field */
                            sortOrder?: number;
                            /** @description A datetime field */
                            createdAt?: string;
                            /** @description A datetime field */
                            updatedAt?: string;
                            /**
                             * @description A datetime field
                             * @default 2026-06-04T06:49:03.662Z
                             */
                            publishedAt: string;
                            /** @description A relational field */
                            category?: components["schemas"]["ApiCategoryCategoryDocument"];
                            /** @description A relational field */
                            product?: components["schemas"]["ApiProductProductDocument"];
                        };
                    };
                };
            };
            /** @description Bad request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Unauthorized */
            401: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Forbidden */
            403: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Not found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Internal server error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
// ...
}